Restructured directories so that css files are stored in their respective Component directory.
Moved Home into pages directory.

Rerouted Home page to '/' route.
