+++ front-end-app/src/MainBody.tsx
@@ -1,4 +1,4 @@
-import Home from "./components/Home";
+import Home from "./components/Pages/Home/Home.tsx";
 import Announcements from "./components/Pages/Announcements/Announcements";
 import DueDates from "./components/Pages/DueDates/DueDates";
 import Events from "./components/Pages/Events/Events";
@@ -7,9 +7,9 @@ import { Route, Routes } from "react-router-dom";
 const MainBody = () => {
   return (
     <Routes>
-      <Route element={<Home />} path="/Home" />
-      <Route element={<Announcements />} path="/Announcements" />
-      <Route element={<Events />} path="/Events" />
+      <Route element={<Home />} path="/" />
+      <Route element={<Announcements />} path="/announcements" />
+      <Route element={<Events />} path="/events" />
       <Route element={<DueDates />} path="/due-dates" />
     </Routes>
   );
