+++ front-end-app/src/components/NavBar/UpperNav/UpperNav.tsx
@@ -9,7 +9,7 @@ const UpperNav = () => {
     const navigate = useNavigate();
 
     const handleClick = () => {
-        navigate('/home');
+        navigate('/');
     }
 
     return (
