+++ front-end-app/src/components/Pages/Announcements/Announcement.css
