+++ front-end-app/src/components/Pages/Announcements/Announcement.tsx
@@ -1,4 +1,4 @@
-import "../../../css/Announcement.css";
+import "./Announcement.css";
 
 interface AnnouncementProps {
   title: string;
