+++ front-end-app/src/components/Pages/DueDates/DueDate.css
@@ -1,5 +1,4 @@
-/* Announcement.css */
-.event {
+.duedate {
   border: 1px solid #ccc;
   padding: 16px;
   margin: 8px 0;
@@ -8,14 +7,19 @@
   position: relative;
 }
 
-.event-header {
+.duedate-header {
   display: flex;
   justify-content: space-between;
-  border-bottom: 1px solid #ccc;
-  padding-bottom: 0px;
+  padding-bottom: 0;
 }
 
-.event-body {
+.duedate-duedate {
+  font-size: 1.4em;
+  color: brown;
+  font-weight: bold;
+}
+
+.duedate-body {
   padding: 4px 0;
   display: flex;
   flex-direction: row;
@@ -23,7 +27,7 @@
   position: relative;
 }
 
-.event-text {
+.duedate-text {
   margin-right: 16px;
   width: 60%;
   word-wrap: break-word;
@@ -31,7 +35,7 @@
   font-size: larger;
 }
 
-.event-image {
+.duedate-image {
   display: flex;
   justify-content: center;
   align-items: center;
@@ -39,13 +43,13 @@
   max-width: none; /* Remove the max-width or adjust as necessary */
 }
 
-.event-image img {
+.duedate-image img {
   max-height: 45vh;
   height: auto; /* Maintains aspect ratio */
   object-fit: cover; /* Ensures the image covers the entire container without stretching */
 }
 
-.event-footer {
+.duedate-footer {
   display: flex;
   justify-content: space-between;
   border-top: 1px solid #ccc;
