+++ front-end-app/src/components/Pages/DueDates/DueDate.tsx
@@ -1,4 +1,4 @@
-import "../../../css/DueDate.css";
+import "./DueDate.css";
 
 interface DueDateProps {
   title: string;
