+++ front-end-app/src/components/Pages/DueDates/DueDates.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import "../../../css/DueDates.css";
+import "./DueDates.css";
 import DueDate from "./DueDate";
 
 interface DueDateData {
