+++ front-end-app/src/components/Pages/Events/Event.css
@@ -1,4 +1,5 @@
-.duedate {
+/* Announcement.css */
+.event {
   border: 1px solid #ccc;
   padding: 16px;
   margin: 8px 0;
@@ -7,19 +8,14 @@
   position: relative;
 }
 
-.duedate-header {
+.event-header {
   display: flex;
   justify-content: space-between;
-  padding-bottom: 0px;
+  border-bottom: 1px solid #ccc;
+  padding-bottom: 0;
 }
 
-.duedate-duedate {
-  font-size: 1.4em;
-  color: brown;
-  font-weight: bold;
-}
-
-.duedate-body {
+.event-body {
   padding: 4px 0;
   display: flex;
   flex-direction: row;
@@ -27,7 +23,7 @@
   position: relative;
 }
 
-.duedate-text {
+.event-text {
   margin-right: 16px;
   width: 60%;
   word-wrap: break-word;
@@ -35,7 +31,7 @@
   font-size: larger;
 }
 
-.duedate-image {
+.event-image {
   display: flex;
   justify-content: center;
   align-items: center;
@@ -43,13 +39,13 @@
   max-width: none; /* Remove the max-width or adjust as necessary */
 }
 
-.duedate-image img {
+.event-image img {
   max-height: 45vh;
   height: auto; /* Maintains aspect ratio */
   object-fit: cover; /* Ensures the image covers the entire container without stretching */
 }
 
-.duedate-footer {
+.event-footer {
   display: flex;
   justify-content: space-between;
   border-top: 1px solid #ccc;
