+++ front-end-app/src/components/Pages/Home/Home.tsx
@@ -1,4 +1,4 @@
-import styles from "../css/Home.module.css";
+import styles from "./Home.module.css";
 
 const Home = () => {
   return (
